+++ src/Akeneo/Pim/Enrichment/AssetManager/Bundle/Resources/public/assets-collection/list.tsx
@@ -36,7 +36,6 @@ import {AssetCollection} from 'akeneopimenrichmentassetmanager/assets-collection
 import {hasValues, isValueComplete} from 'akeneopimenrichmentassetmanager/enrich/domain/model/product';
 import {Family} from 'akeneoassetmanager/platform/model/structure/family';
 import {NoDataSection, NoDataTitle} from 'akeneoassetmanager/platform/component/common/no-data';
-import {RuleRelation} from 'akeneoassetmanager/platform/model/structure/rule-relation';
 import {RuleNotification} from 'akeneoassetmanager/platform/component/rule-notification';
 import {selectErrors} from 'akeneopimenrichmentassetmanager/assets-collection/reducer/errors';
 import {ValidationError} from 'akeneoassetmanager/platform/model/validation-error';
@@ -58,6 +57,7 @@ import {
 } from 'akeneoassetmanager/platform/model/structure/attribute-group';
 import {AssetsIllustration, Information, Link, LockIcon} from 'akeneo-design-system';
 import {ButtonContainer} from 'akeneoassetmanager/application/component/app/button';
+import {RulesNumberByAttribute} from 'akeneoassetmanager/platform/model/structure/rule-relation';
 
 type ListStateProps = {
   attributes: Attribute[];
@@ -67,7 +67,7 @@ type ListStateProps = {
   productLabels: LabelCollection;
   family: Family | null;
   context: Context;
-  ruleRelations: RuleRelation[];
+  rulesNumberByAttribute: RulesNumberByAttribute;
   errors: ValidationError[];
 };
 type ListDispatchProps = {
@@ -81,7 +81,7 @@ type DisplayValuesProps = {
   productLabels: LabelCollection;
   family: Family | null;
   context: Context;
-  ruleRelations: RuleRelation[];
+  rulesNumberByAttribute: RulesNumberByAttribute;
   errors: ValidationError[];
   onChange: (value: Value) => void;
 };
@@ -139,7 +139,7 @@ const DisplayValues = ({
   attributeGroups,
   family,
   context,
-  ruleRelations,
+  rulesNumberByAttribute,
   onChange,
   errors,
   productIdentifier,
@@ -208,7 +208,7 @@ const DisplayValues = ({
               </ButtonContainer>
             )}
           </SectionTitle>
-          <RuleNotification attributeCode={value.attribute.code} ruleRelations={ruleRelations} />
+          <RuleNotification attributeCode={value.attribute.code} rulesNumberByAttribute={rulesNumberByAttribute} />
           <ValidationErrorCollection attributeCode={value.attribute.code} context={context} errors={errors} />
           <AssetCollection
             productIdentifier={productIdentifier}
@@ -229,7 +229,7 @@ const List = ({
   attributeGroups,
   family,
   context,
-  ruleRelations,
+  rulesNumberByAttribute,
   errors,
   productIdentifier,
   productLabels,
@@ -242,7 +242,7 @@ const List = ({
         attributeGroups={attributeGroups}
         family={family}
         context={context}
-        ruleRelations={ruleRelations}
+        rulesNumberByAttribute={rulesNumberByAttribute}
         onChange={onChange}
         errors={errors}
         productIdentifier={productIdentifier}
@@ -277,7 +277,7 @@ export default connect(
     productIdentifier: selectProductIdentifer(state),
     productLabels: selectProductLabels(state),
     family: selectFamily(state),
-    ruleRelations: selectRuleRelations(state),
+    rulesNumberByAttribute: selectRuleRelations(state),
     errors: selectErrors(state),
   }),
   (dispatch: any): ListDispatchProps => ({