+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 - PIM-9676: Reference Entities - Fix missing warning message when changing page via the breadcrumb with unsaved changes.
 - PIM-9668: Asset attribute - Fix Text area + Rich text editor modes turn text attributes into infinite extendable fields when not using spaces
 - PIM-9664: Display Ziggy as asset image when the preview cannot be generated
+- PIM-9702: Fix infinite loop when using get all assets API endpoint
+- PIM-9720: Fix Asset cursor search after query
+- PIM-9723: Fix Mysql memory size issue during ordering in SQL when getting category trees in product grid
+- PIM-9722: Fix the increasing amount of requests when editing attribute with options
 
 ## Improvements